Dosbox Pure: revert to old version 0.21 (+keep the new one 0.9.7)
Dosbox 0.21 seems faster and easier to configure.
We keep Dosbox 0.9.7 for users who want to benefits to the fixes & improvements of this version (e.g. bypass the limitation of 16 keyboard bindings with mouse mode).

So it will be possible to switch the default core to v0.9.7 thank to GLO (Game List Options)

Feedback on 0.9.7 here :
https://github.com/OnionUI/Onion/issues/611

And here :
https://discord.com/channels/529983248114122762/931367023588569180/1042760981668962364