Update Gambatte with new color management (981096d)
This is a special core which allows to navigate in a better way between the palettes.
Source :
https://github.com/schmurtzm/gambatte-libretro/tree/advanced-color-organisation
(commit 981096d)
For similar palettes, the best one has been chosen based on both readability and color ramping. All categories have been organized to progress through the color spectrum as best as can be managed starting with the color green.

For nostalgic people, the majority of the old palettes are in the "Extra" and "Others" categories, so if you miss one you'll probably find it here.