Major improvements to mp4_to_mng.ps1
Major update. The script now runs recursively for all systems at once. It also removes the first 17 frames to avoid black fade in, removing the need to also run remove10.ps1.

It also removes the same time period from the mp3 to keep the mng and mp3 in sync. Not only that, but it also puts the filesin the proper Snaps folder locations.

Not to mention that it looks much nicer while running.