Replacing unzip with 7z in installer
7z is more reliable than unzip from the busybox.
The memory consumption and reliability for extracting large files could be better.
The 7z command line use -bsp1 -bb to force the % progression in a log file called extraction_output.
Then extraction_status is parsed to transfer the extracting progression to installUI.
As the extracting command runs in another shell the file called "extraction_status" will get the result of the 7z command.