Add test code
There's currently an issue, where if runtime.sh starts the process it all works fine, but if tweaks starts it, it does not. (This seems wrong has runtime.sh is the parent of tweaks).

The current code on this branch requires you to:

Open tweaks
Turn samba on
Exit tweaks

This causes a naturally start to Samba.

If you remove start_smbd from check function, and move line 135 down into check_smbdstate it causes a situation where smbd will start but the shares aren't browsable.

If you add a call to "check" function in "check_smbdstate" it also doesn't work.. the call to start needs to come from runtime.sh

It's either an env var issue or a permissions issue.