FIX: MM showing MY354 instead of MY283 (#989)
Currently, the MainUI binaries for the 283 show MY354 in the
Settings/About device screen.
Fixed by patching the 283-clean/283-expert binaries.

Additionaly, mount MainUI instead of copying it. This should reduce wear
of the SD card.
It might also fix the bug where the wrong MainUI would be loaded ("No
Wifi option on MM+")
since there is a lot less of file checking going on.

![MainUI_000](https://github.com/OnionUI/Onion/assets/6023076/577e24db-7502-4a1e-b83b-d8604b425dcd)

---------

Co-authored-by: Aemiii91 <44569252+Aemiii91@users.noreply.github.com>