Remove reliance on the .auth file for telnetd
telnetd now uses any user from the new passwd file (provided it's mounted) - however to note; telnetd will run the /etc/profile script again when you login - we should consider mounting this script and adding a tracking file somewhere (a .run_script_once check). It's harmless when logged in as a none root user, but when logged in as root it will cause problems.