FIX: Swap start and select button in MainUI Gamelist (#1006)
Patched MainUI binaries to swap start and select button in the game
list.
This fixes #995 

In my testing, this only affected the game list. For example, pressing
select on the main menu or emulator list still brings up the "Refresh
all roms" menu

Test:
for all 4 binaries:

- [x] On main menu, press select, observe the "refresh all roms" menu
pops up
- [x] In the emulator list, press select, observe the "refresh all roms"
menu pops up
- [x] Open an emulator, press select. Observe the menu pops up, but the
Emulator name doesn't change to the rom path
- [x] Close the menu and press start. Observe that the Emulator name
changes to the rom path (and is displayed on the left)
- [x] Go to "Favorites" and press select. Observe that it still brings
up the popup menu.
- [x] Go to "Apps" and press select. Observe that it still brings up the
popup menu.

Co-authored-by: Aemiii91 <44569252+Aemiii91@users.noreply.github.com>