FEAT: Added tool to automatically generate miyoogamelists with clean names (#1039)
Add a script/tool that generate `miyoogamelist.xml` for all systems
(except Arcades).
I chose to exclude arcade titles because they have their own gamelist
already.

In the process, it removes additional info from the name of the game,
assuming the roms are names according to [No-Intro Naming
Convention](https://wiki.no-intro.org/index.php?title=Naming_Convention).

Also, the script will backup existing `miyoogamelist.xml` files to
`miyoogamelist.xml.bak` in case the user add one already and avoid
losing it.

~~There is room for improvement, for instance replacing " - " by ": " in
names or moving back articles to the beginning o the name (ie: "Legend
of Zelda, The" -> "The Legend of Zelda")~~
Edit: Articles (The, A, An) are moved to the start of the name
(according to naming convention) and subtitles are separated using '": "
instead of " - "