FIX: Fix cleanup crash - Netplay client scripts (#1059)
## Netplay client

- `Fixed cleanup loop error` (and potentially rare upside down screen
error) caused by an unintended loop in cleanup()

- `Refactored some functions to cut down on bloat in wpa_supp.conf` (in
some scenarios wpa_cli is unavailable to us to use when wpa_supp pid
isn't running and ctrl int doesn't exist so we can't remove the AP entry
we made) Stop bloating wpa_supplicant in cases where cleanup would fail
to reconfigure (cleanup **must** still run! every time!)

- `Add fallback for IP address assignment` (pick a random one in range
if we weren't successfull getting one from DHCP) - Connecting from a
`wifi down` stage has trouble with DHCP at times - improve UX with a one
shot connection.

- `Remove artificial sleeps from cleanup as no longer required` - Cut
down on cleanup time on clients

- `Remove some infopanels that aren't beneficial for the user to see` -
No point user seeing that the core/rom matches, if the game starts it
does, if not they'll see the error.

## Pokemon client

- Same logic as above, aiming to cut down on bloat in wpa_supp, fix
issues with joining when wifi is disabled and fix re-association to
default network when wifi turns back on.

- `Pushed delays in connection stage out further` - catch clients being
slow at joining the hotspot and getting connectivity
- `Push some delays out for save sync` - Retroarch can sometimes lock up
on the host (cause unknown) when exiting, it does unlock but we blow
past the timer on the client so save sync fails.. This tends to happen
for shorter sessions like what you'd see in testing, but full sessions
don't seem to have this issue.