FIX: Add migration script for hostapd.conf fix (#1113)
Minor change.
## What:
PR to add a migration script to add the missing line to the top of the
.conf file.
## Why:
Fixes "Hostapd hook failing" loop when trying to setup an easy netplay
session for pokemon
## When:
Hostapd ctrl_interface was added in ENP PR to the conf, the conf won't
get updated without a full install.

![image](https://github.com/OnionUI/Onion/assets/47260768/88669532-7b5e-4c0a-90cd-66c4b941e094)


![image](https://github.com/OnionUI/Onion/assets/47260768/b9129245-17bf-46af-9dbd-4b901853e7b8)