Fix exit on Scraper
If you modify some options and then launch a scraping, Scraper doesn't exit at the end of the scraping process and displays options menu again.
To avoid that we use pkill to systematically exit ST (terminal app) at the end of the scraping process.

A more precise analysis will have to be carried out to determine the origin of this bug, but in the meantime this workaround will do the trick.