Fix Easy Netplay with other themes than Silky
It seems that with some themes infoPanel --persistent is not dismissed.
Probably due to the fact that infoPanel is launched in background : a question of synchronization with the flag (the previous infoPanel erase the flag of the next infoPanel ?)

These modifications with sync commands and a pkill should definitively solve the problem.