FIX: Netplay improvements (#1242)
It seems that with some themes infoPanel --persistent is not dismissed.
Probably due to the fact that infoPanel is launched in background : a
question of synchronization with the flag (the previous infoPanel erase
the flag of the next infoPanel ?)

These modifications with sync commands and a pkill should definitively
solve the problem.
**Fixes :** 

- Disable retroachievements when Netplay is used
- Fix Easy Netplay with other themes than Silky

> _It seems that with some themes infoPanel --persistent is not
dismissed.
> Probably due to the fact that infoPanel is launched in background : a
question of synchronization with the flag (the previous infoPanel erase
the flag of the next infoPanel ?)
> These modifications with sync commands and a pkill should definitively
solve the problem._

- Adding sync after each "dismiss_info_panel" flag creation


**Update :** 

- Completing Netplay core compatibility file : _NEOGEO, PCE and CPS1
added + alphabetical reordering_

**Other fixes :** 
- Fix region settings for screenscraper