FEAT: PWM control (#1179)
Adds the option to control the backlights PWM frequency.

This can improve the constant buzzing of the device by making it a less
annoying frequency.
From my testing, at 200 Hz the buzzing is much less noticeable and there
does not seem to be a negative impact on the image quality. At 100 Hz
the flicker might be visible to sensitive eyes.

I would have preferred to go above the audible hearing range, but
everything above 1000 Hz makes the brightness control unstable/unusable.

MainUI has been patched to not set the PWM value every time it is
launched.

---------

Co-authored-by: tGecko <tGecko@users.noreply.github.com>
Co-authored-by: Schmurtz <Schmurtz@gmail.com>
Co-authored-by: schmurtzm <schmurtzm@users.noreply.github.com>