Fix the sighandler in MainUI (#1193)
MainUI patched to react to SIGTERM signal.

A sighandler to break out of the main loop was incorrectly implemented,
set up to act on SIGKILL (9) which is not possible, as SIGKILL will
always kill and leave the process no time to react.


![image](https://github.com/OnionUI/Onion/assets/6023076/8934189a-fe33-4dd6-97fc-25fd4c0e5b3a)

The sighandler simply sets a value which is checked early in the main
`while(true)` loop

This allows us to shut down MainUI gracefully by sending SIGTERM instead
of SIGKILL which can cause all sorts of troubles, notably black screens.

This _includes_ the patches from #1179 and thus should be merged after
#1179

---------

Co-authored-by: Schmurtz <Schmurtz@gmail.com>