Indicate to user if they've ignored the schedule by using hotkey
The logic behind this is:

If the user disables the BLF for some reason while schedule is enabled we'll skip the schedule for this session, the user can remove this flag (/tmp/.blfIgnoreSchedule) by flipping the schedule on/off.

I chose this method of wiping the schedule flag so if the user takes no action, when they restart the schedule is still on for the next session.