Norwegian language cleanups (#1332)
I've been using the translation that's currently in the main branch for
a little bit and noticed some small typos and inconsistencies that I
have corrected. I also made the change of renaming the FC and SFC
systems to NES and SNES as these are the names the systems were released
under in Norway but I'm not married to that particular change.

Co-authored-by: Schmurtz <schmurtz@gmail.com>