FEAT: Support for 752x560 resolution used by new MMv4 screens (#1272)
Support for 752x560 resolution.

On startup, runtime checks the physical screen resolution and stores it
to `/tmp/screen_resolution`
If the firmware is 202310271401 and the physical resolution is 752x560,
it also creates the flag `/tmp/new_res_available`

If this flag is set on game launch, the resolution is changed and keymon
and batmon are informed of this via signal so they can update their
stored resolutions.

Keymon was updated to respect the render resolution when drawing the
brightness bar, batmon was updated to respect the render resolution when
drawing the "battery empty" icon.

TODO:

- [x] implement either a blacklist for anything hardcoded to 640x480 or
a whitelist for anything that supports 752x560?
    so far only drastic is added as exception.

- [x] testing, finding edge cases, other resolution dependent things?
- [ ] overlays

---------

Co-authored-by: Schmurtz <schmurtz@gmail.com>
Co-authored-by: schmurtzm <schmurtzm@users.noreply.github.com>