FIX: BLF Menu lockup in certain circumstances (#1337)
If you get segfaults with this PR, run this in terminal before (The max
value is no longer 5):

```rm -rf /mnt/SDCARD/.tmp_update/config/display/blueLightLevel```

- Remove "None" - it's redundant and causes issues
- Fix menu lockup when scrolling, check if BLF is actually on before setting "blf_changing = true;" to negate needlessly locking the menu items. The artificial lockup on menu items is only needed when BLF is actually changing to avoid race conditions and pulling the display settings out of sync (causes flickering/bad previous state of RGB values, etc)

Fixes: https://github.com/OnionUI/Onion/commit/ddc9a9a9bfb45655e3fad3457f4145ae7d8e5dc3#commitcomment-135316788

This is only a fix for when in the tweaks menu. The artificial delay is required. Setting display values into the screen gives no easy way to read them besides saving to file (this feature runs inside and outside tweaks so it's a settings value that's bitshifted). 

Ideally this feature should run on a command queue/buffer and be processed as they come in so we don't have to try and keep sync...

It would also be great to get a "spinner" as discussed in discord, for when we don't need a full dialog, but need to show the user that something is happening in the background.. Something to look at in tweaks for the future i think.

---------

Co-authored-by: XK9274 <XK9274@users.noreply.github.com>
Co-authored-by: Schmurtz <schmurtz@gmail.com>