FIX: Replace global resolution blacklist with a per-game/app whitelist file (#1346)
Replace the central res_exceptions file, which had a list of strings run
against cmd_to_run.sh, with a per-program approach.
If a game/app wants to launch in 560p, it needs a file `full_resolution`
next to it's launch.sh.

This has multiple benefits, for example

- No need to iterate over a growing file on each game/app launch
- a user can easily remove or add the file if needed

A drawback would be that this is more of a whitelist approach, so for
every system/app/game which can run in 560p, the according file must be
created.
This is done at build time by the newly added script
[create_fullres_files.sh](https://github.com/OnionUI/Onion/compare/replace-blacklist-with-something-better#diff-bca2ea91900785aa120dd3ba50be7f10f914e7b43acd0f6600b6a6a66df2fec6)