Fix lockups entering apps or games  (#1369)
# TODO: check anywhere else we may need to now set the TZ... 

We've found this is caused by a weird config combination... where if a
certain TZ is set in runtime, MainUI inherits it and it causes some
loops/threads within MainUI to not break out (it gets stuck in a
threaded loop which will never exit), so you get stuck at a "Loading"
screen when entering Roms, or a complete freeze when entering Apps.

It could be a mismatch of TZ between MainUI and everything that runs
"before" it.. IE: updater, customer/main etc. & The kernel will never
know about the TZ as we can't insert it so early. This is speculation,
the root cause i haven't found.

This is obvious when files are removed from the SD and the time/date of
creation checked on a windows PC, the TZ metadata is sometimes bad.

This pr will:
- Remove runtime.sh environments knowledge of the timezone, therefore
MainUI no longer knows about the TZ either.
- Pass cmd_to_run.sh the timezone for apps and games to be aware it's
potentially changed, and to set theirs correctly
- Tell BLF script about the TZ manually, as keymon now no longer knows
the TZ.

TZ is already set in update_networking.sh main function so no changes
needed there for anything called as they will inherit this TZ from their
parent.

Games and apps will inherit this from pushing it through cmd_to_run.sh

To verify this PR:

- Rename your .tmp_update/config directoy
- Drop in the attached tainted config dir:
[config.zip](https://github.com/OnionUI/Onion/files/13928317/config.zip)
- Open the search app with something in the search field
- Hard power down with the power button
- Power on & try and load an app or game -> freeze.

Add in the new runtime.sh from this PR and re-run the above steps,
observe no freeze.

The TZ is still correctly passed to retroarch:

![image](https://github.com/OnionUI/Onion/assets/47260768/546f5398-acfd-42ae-819a-4bbf3cf0b791)

And any apps:

![image](https://github.com/OnionUI/Onion/assets/47260768/5044e025-7cd6-4782-bebb-11d75021c0e0)

Aslong as they're a child of either update_networking or cmd_to_run.sh

---------

Co-authored-by: Aemiii91 <44569252+Aemiii91@users.noreply.github.com>
Co-authored-by: Totofaki <16885275+Sichroteph@users.noreply.github.com>
Co-authored-by: Schmurtz <schmurtzm@users.noreply.github.com>
Co-authored-by: tGecko <tGecko@proton.me>