Fix infoPanel crashes when loading incompatible images, that also need to be stretched (#1418)
Closes #1386 

The images that crashed the infoPanel were in a different format, which
SDL_SoftStretch couldn't handle, leaving image_to_draw empty and
subsequently crashing when drawing the empty image.
To safeguard against this, create a separate surface and check the
return of SoftStretch for error.

Testing this showed that some of the stretched images looked horrible,
so now only stretch images if they are too big to fit the screen. (MMv4
752x560 screenshots mainly)

Additionally, add some error handling to the Gallery launch script, so
it would show "Error" instead of "No screenshots found" when infoPanel
crashes to be less misleading.

To test, copy the screenshot folder from the issue #1386 to your
SD/Screenshots and confirm it doesn't crash when loading the different
screenshots.

---------

Co-authored-by: XK <47260768+XK9274@users.noreply.github.com>
Co-authored-by: Aemiii91 <44569252+Aemiii91@users.noreply.github.com>