Fix: broken link in advance guides (#1489)
Updated Big Zoom link to my server hosting the file here:
https://files.phyrextechs.com/share/BSKak5jy

Not sure if this is frowned upon here and it would be better to host
somewhere else. Please let me know.

Sorry for making another PR I'm new and was unsure of the proper
protocol and messed up my last branch.

Please direct me to any information on the proper ways to help out as I
don't want to create more work for others.

Thanks.

---------

Co-authored-by: Aemiii91 <44569252+Aemiii91@users.noreply.github.com>