Move writing settings to thread. (#1488)
I had a problem with volume control. I think it was because SD card is
not a fast enough. And writing settings are executes synchrony. I tryed
to move it into another thread.

How it works:
When settings changed we are start writing to SD card asynchrony instead
of synchrony. If new changes try to run cycle again it will be delayed
until thread not finished.
Then settings will be writen one more time. It can be not a thread safe,
when main thread re-write settings variables and the second thread try
to read them for store to SD card.
In all cases, even if writen to SD card settigns will be changed in
"writing" moment, it will be re-writen one more time.

Video how it works before with slow SD card and how it works after
moving saving settings to thread:
https://www.youtube.com/watch?v=fIy-bYyZLBk

---------

Co-authored-by: andrei nikulin <nikulin.aa@kvanta.pro>