Implement RTC detection for devices that have it (newer MM+) (#1665)
Since the newest hardware revisions of the Miyoo Mini+ have RTC, it
should be handled accordingly.

Detects if RTC is present, and does not restore and/or advance time if
it is.
Works with newest Miyoo Mini+ with built in RTC as well as modded in
RTC.

Edit: will also work on MM with RTC, see #1673