Fixed miyoo mini inaccurate battery readings. (#1712)
This PR fixes the following issues on the Miyoo Mini:
- Fixes sudden low battery warnings or shutdowns do to
unstable/unsmoothed voltage reading
- Fixes inaccurate battery readings
- Fixes inaccurate voltage droop curve
- Prevent battery rising again due to system power draw changes. (about
80mv variance, a single percent on the battery)
- Fixes device not cutting of for 40 minutes to 1hr after hitting 0%
battery

The existing issue was due to reading ADC value as if it was 4.3V to
3.3V.
The correct range should be from 4.2V to 3.0V as the miyoomini has a
li-ion, not a li-po battery.
Additionally I havee simplified the ADC to voltage reading, and
interpolated the voltage across a lookup table to make the battery
percentage relation completly linear.

I also collected 10 samples of the battery reading and averaged them
together to better smooth the adc data.

These changes have completely fixed the wildly inaccurate and strange
battery readings for the Miyoo Mini.
And they have made the battery drain curve completely linear, with
proper power cutoff right at 0%.

If any one has a miyoo mini also, please give it a go :)


![image](https://github.com/user-attachments/assets/646f63f1-dfe1-446c-b2d7-2ee5a5ec532b)

---------

Co-authored-by: Aemiii91 <44569252+Aemiii91@users.noreply.github.com>